Refactoring of functions into classes (#3)
* moved functions calls behavior to separate classes

* bumped version
