+++ src/Scrima.OData.AspNetCore/Scrima.OData.AspNetCore.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFrameworks>netcoreapp3.1;net5.0</TargetFrameworks>
-    <Version>1.2.2</Version>
+    <Version>1.2.3</Version>
     <Title>Scrima.NET OData parsing for AspNet.Core</Title>
   </PropertyGroup>
