+++ src/Scrima.Queryable/Functions/TrimFunction.cs
@@ -1,0 +1,22 @@
+﻿using System.Collections.Generic;
+using System.Linq.Expressions;
+
+namespace Scrima.Queryable.Functions
+{
+    internal class TrimFunction : ScrimaQueryableFunction
+    {
+        public override string FunctionName => "trim";
+
+        public override Expression CreateExpression(IList<Expression> arguments)
+        {
+            ValidateParameterCount(arguments, 1);
+
+            if (arguments[0].Type != TypeUtilities.StringType)
+            {
+                return InvalidParameterTypes("strings");
+            }
+
+            return Expression.Call(arguments[0], Methods.StringTrim);
+        }
+    }
+}
