+++ src/Scrima.Queryable/Scrima.Queryable.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
-    <Version>1.2.2</Version>
+    <Version>1.2.3</Version>
     <Title>Scrima.NET execution engine for IQueryable sources</Title>
   </PropertyGroup>
